points to db

+++ ACS_Trend/Domain/Entities/TrendPoint.cs
@@ -14,8 +14,8 @@ namespace ACS_Trend.Domain.Entities
         public int ID_TrendPoint { get; set; }
         [ForeignKey("Trend")]
         public int TP_ID_Trend { get; set; }
-        public float Date_time { get; set; } // DateTime
-        public float Parameter { get; set; }
+        public double Date_time { get; set; } // DateTime
+        public double Parameter { get; set; }
 
 
         public virtual Trend Trend { get; set; }
